Feat: add both form and form validation
This form has a function that validates the user data. It display
several errors, one at a time. For the moment, it does nothing when
the user submits valid data.

The form validation is located in a js file, this is done to prevent
the App.jsx file getting larger.

Also created style files for the form sections, but these files doesn't
contain rules at the moment of writing this.