Feat: let the user submit and edit cv
Now when the user submits a cv, it shows the info with html elements
instead of the form, along with a button that edits the information and
brings back the form, but now with the input fields filled with the
previous values.

There are two new files, one for the form and another for the cv with
html elements.

The previous component files (GeneralInformation, EducationalExperience
and PracticalExperience) now receive props. These props are the formData
state variable and its function setFormData. This, along with a onChange
method, makes the input fields controlled components, updating its
values and storing them for further validation and showing them when
they're valid.

Removed the e parameter in validateForm of form-validation.js because
it's no longer needed. Now the form in CVForm has a function that
receives the event parameter to prevent form submission and calls this
validateForm function.